django/requirements.txt
django==5.2.2
django-ninja==1.3.0
django-allauth[socialaccount,headless,idp-oidc]==65.15.0
django-cors-headers==4.7.0
django-storages==1.14.4
django-ratelimit==4.1.0
django-easy-audit==1.3.7
django-solo==2.4.0
django-celery-results==2.6.0
django-celery-beat==2.8.1
dj-database-url==2.3.0
dj-email-url==1.0.6
gunicorn==23.0.0
celery==5.5.3
redis==7.0.1
hydroserver-sensorthings==0.4.2
pandas==2.2.3
orjson==3.10.15
uuid6==2024.7.10
croniter==6.0.0
python-decouple==3.8
pyyaml==6.0.2
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
country-list==1.1.0
hsclient==1.1.6
boto3==1.36.15
google-cloud-storage==3.0.0
